Split checksum generation out
